use f32 instead of f64 in route splines; saves a ton of memory

+++ engine/route/query.rs
@@ -81,7 +81,7 @@ fn construct_route(
             distance: end_dist,
         }))
         .collect(),
-        cost,
+        cost as f32,
         input,
     )
 }
